Move reference-transform/fillet helpers to helpers/referenceTransforms.ts
Also moves transformStlFeatureData into helpers/transform.ts (needed by
reference transforms without creating a cycle back to projectStore).
Slices and external consumers now import these directly.
